agent: start to rework the debug console
It's the first commit of the rework.

Fixes: #1647
